Prevents any HTTP queries from triggering exceptions. Creates AccountTable to be a good home for all the code relating to accounts and Tokens.
